backend/requirements.txt
fastapi==0.115.5
uvicorn==0.32.1
pydantic==2.10.2
llama_index==0.12.2
nest_asyncio==1.6.0
wikipedia_api==0.7.1
python_dotenv==0.21.0
